+++ docs/source/conf.py
@@ -84,9 +84,11 @@ extensions = [
     'local_customization',
 ]
 
+# FIXME: change the "python" link back to /3 when Python 3.11 is released
 intersphinx_mapping = {
-    "python": ('https://docs.python.org/3', None),
+    "python": ('https://docs.python.org/3.11', None),
     "outcome": ('https://outcome.readthedocs.io/en/latest/', None),
+    "pyopenssl": ('https://www.pyopenssl.org/en/stable/', None),
 }
 
 autodoc_member_order = "bysource"