+++ trio/_core/_multierror.py
@@ -1,9 +1,15 @@
 import sys
-import traceback
-import textwrap
 import warnings
+from typing import Sequence
 
 import attr
+
+from trio._deprecate import warn_deprecated
+
+if sys.version_info < (3, 11):
+    from exceptiongroup import BaseExceptionGroup, ExceptionGroup, print_exception
+else:
+    from traceback import print_exception
 
 ################################################################
 # MultiError
@@ -127,7 +133,7 @@ class MultiErrorCatcher:
 
     def __exit__(self, etype, exc, tb):
         if exc is not None:
-            filtered_exc = MultiError.filter(self._handler, exc)
+            filtered_exc = _filter_impl(self._handler, exc)
 
             if filtered_exc is exc:
                 # Let the interpreter re-raise it
@@ -151,7 +157,7 @@ class MultiErrorCatcher:
                 del _, filtered_exc, value
 
 
-class MultiError(BaseException):
+class MultiError(BaseExceptionGroup):
     """An exception that contains other exceptions; also known as an
     "inception".
 
@@ -174,21 +180,23 @@ class MultiError(BaseException):
 
     """
 
-    def __init__(self, exceptions):
-        # Avoid recursion when exceptions[0] returned by __new__() happens
-        # to be a MultiError and subsequently __init__() is called.
-        if hasattr(self, "exceptions"):
-            # __init__ was already called on this object
-            assert len(exceptions) == 1 and exceptions[0] is self
-            return
-        self.exceptions = exceptions
+    def __init__(self, exceptions, *, _collapse=True):
+        self.collapse = _collapse
 
-    def __new__(cls, exceptions):
+        # Avoid double initialization when _collapse is True and exceptions[0] returned
+        # by __new__() happens to be a MultiError and subsequently __init__() is called.
+        if _collapse and getattr(self, "exceptions", None) is not None:
+            # This exception was already initialized.
+            return
+
+        super().__init__("multiple tasks failed", exceptions)
+
+    def __new__(cls, exceptions, *, _collapse=True):
         exceptions = list(exceptions)
         for exc in exceptions:
             if not isinstance(exc, BaseException):
-                raise TypeError("Expected an exception object, not {!r}".format(exc))
-        if len(exceptions) == 1:
+                raise TypeError(f"Expected an exception object, not {exc!r}")
+        if _collapse and len(exceptions) == 1:
             # If this lone object happens to itself be a MultiError, then
             # Python will implicitly call our __init__ on it again.  See
             # special handling in __init__.
@@ -200,13 +208,23 @@ class MultiError(BaseException):
             # In an earlier version of the code, we didn't define __init__ and
             # simply set the `exceptions` attribute directly on the new object.
             # However, linters expect attributes to be initialized in __init__.
-            return BaseException.__new__(cls, exceptions)
+            if all(isinstance(exc, Exception) for exc in exceptions):
+                cls = NonBaseMultiError
+
+            return super().__new__(cls, "multiple tasks failed", exceptions)
 
     def __str__(self):
         return ", ".join(repr(exc) for exc in self.exceptions)
 
     def __repr__(self):
         return "<MultiError: {}>".format(self)
+
+    def derive(self, __excs):
+        # We use _collapse=False here to get ExceptionGroup semantics, since derive()
+        # is part of the PEP 654 API
+        exc = MultiError(__excs, _collapse=False)
+        exc.collapse = self.collapse
+        return exc
 
     @classmethod
     def filter(cls, handler, root_exc):
@@ -224,7 +242,12 @@ class MultiError(BaseException):
           ``handler`` returned None for all the inputs, returns None.
 
         """
-
+        warn_deprecated(
+            "MultiError.filter()",
+            "0.22.0",
+            instead="BaseExceptionGroup.split()",
+            issue=2211,
+        )
         return _filter_impl(handler, root_exc)
 
     @classmethod
@@ -236,12 +259,23 @@ class MultiError(BaseException):
           handler: as for :meth:`filter`
 
         """
+        warn_deprecated(
+            "MultiError.catch",
+            "0.22.0",
+            instead="except* or exceptiongroup.catch()",
+            issue=2211,
+        )
 
         return MultiErrorCatcher(handler)
 
 
+class NonBaseMultiError(MultiError, ExceptionGroup):
+    pass
+
+
 # Clean up exception printing:
 MultiError.__module__ = "trio"
+NonBaseMultiError.__module__ = "trio"
 
 ################################################################
 # concat_tb
@@ -361,95 +395,8 @@ def concat_tb(head, tail):
     return current_head
 
 
-################################################################
-# MultiError traceback formatting
-#
-# What follows is terrible, terrible monkey patching of
-# traceback.TracebackException to add support for handling
-# MultiErrors
-################################################################
-
-traceback_exception_original_init = traceback.TracebackException.__init__
-
-
-def traceback_exception_init(
-    self,
-    exc_type,
-    exc_value,
-    exc_traceback,
-    *,
-    limit=None,
-    lookup_lines=True,
-    capture_locals=False,
-    compact=False,
-    _seen=None,
-    **kwargs,
-):
-    if sys.version_info >= (3, 10):
-        kwargs["compact"] = compact
-
-    # Capture the original exception and its cause and context as TracebackExceptions
-    traceback_exception_original_init(
-        self,
-        exc_type,
-        exc_value,
-        exc_traceback,
-        limit=limit,
-        lookup_lines=lookup_lines,
-        capture_locals=capture_locals,
-        _seen=_seen,
-        **kwargs,
-    )
-
-    seen_was_none = _seen is None
-
-    if _seen is None:
-        _seen = set()
-
-    # Capture each of the exceptions in the MultiError along with each of their causes and contexts
-    if isinstance(exc_value, MultiError):
-        embedded = []
-        for exc in exc_value.exceptions:
-            if id(exc) not in _seen:
-                embedded.append(
-                    traceback.TracebackException.from_exception(
-                        exc,
-                        limit=limit,
-                        lookup_lines=lookup_lines,
-                        capture_locals=capture_locals,
-                        # copy the set of _seen exceptions so that duplicates
-                        # shared between sub-exceptions are not omitted
-                        _seen=None if seen_was_none else set(_seen),
-                        **kwargs,
-                    )
-                )
-        self.embedded = embedded
-    else:
-        self.embedded = []
-
-
-traceback.TracebackException.__init__ = traceback_exception_init  # type: ignore
-traceback_exception_original_format = traceback.TracebackException.format
-
-
-def traceback_exception_format(self, *, chain=True):
-    yield from traceback_exception_original_format(self, chain=chain)
-
-    for i, exc in enumerate(self.embedded):
-        yield "\nDetails of embedded exception {}:\n\n".format(i + 1)
-        yield from (textwrap.indent(line, " " * 2) for line in exc.format(chain=chain))
-
-
-traceback.TracebackException.format = traceback_exception_format  # type: ignore
-
-
-def trio_excepthook(etype, value, tb):
-    for chunk in traceback.format_exception(etype, value, tb):
-        sys.stderr.write(chunk)
-
-
-monkeypatched_or_warned = False
-
+# Remove when IPython gains support for exception groups
+# (https://github.com/ipython/ipython/issues/13753)
 if "IPython" in sys.modules:
     import IPython
 
@@ -459,24 +406,19 @@ if "IPython" in sys.modules:
             warnings.warn(
                 "IPython detected, but you already have a custom exception "
                 "handler installed. I'll skip installing Trio's custom "
-                "handler, but this means MultiErrors will not show full "
+                "handler, but this means exception groups will not show full "
                 "tracebacks.",
                 category=RuntimeWarning,
             )
-            monkeypatched_or_warned = True
         else:
 
             def trio_show_traceback(self, etype, value, tb, tb_offset=None):
                 # XX it would be better to integrate with IPython's fancy
                 # exception formatting stuff (and not ignore tb_offset)
-                trio_excepthook(etype, value, tb)
+                print_exception(value)
 
-            ip.set_custom_exc((MultiError,), trio_show_traceback)
-            monkeypatched_or_warned = True
+            ip.set_custom_exc((BaseExceptionGroup,), trio_show_traceback)
 
-if sys.excepthook is sys.__excepthook__:
-    sys.excepthook = trio_excepthook
-    monkeypatched_or_warned = True
 
 # Ubuntu's system Python has a sitecustomize.py file that import
 # apport_python_hook and replaces sys.excepthook.
@@ -490,27 +432,21 @@ if sys.excepthook is sys.__excepthook__:
 # hook.
 #
 # More details: https://github.com/python-trio/trio/issues/1065
-if getattr(sys.excepthook, "__name__", None) == "apport_excepthook":
+if (
+    sys.version_info < (3, 11)
+    and getattr(sys.excepthook, "__name__", None) == "apport_excepthook"
+):
+    from types import ModuleType
+
     import apport_python_hook
+    from exceptiongroup import format_exception
 
     assert sys.excepthook is apport_python_hook.apport_excepthook
 
-    # Give it a descriptive name as a hint for anyone who's stuck trying to
-    # debug this mess later.
-    class TrioFakeSysModuleForApport:
-        pass
+    def replacement_excepthook(etype, value, tb):
+        sys.stderr.write("".join(format_exception(etype, value, tb)))
 
-    fake_sys = TrioFakeSysModuleForApport()
+    fake_sys = ModuleType("trio_fake_sys")
     fake_sys.__dict__.update(sys.__dict__)
-    fake_sys.__excepthook__ = trio_excepthook  # type: ignore
+    fake_sys.__excepthook__ = replacement_excepthook  # type: ignore
     apport_python_hook.sys = fake_sys
-
-    monkeypatched_or_warned = True
-
-if not monkeypatched_or_warned:
-    warnings.warn(
-        "You seem to already have a custom sys.excepthook handler "
-        "installed. I'll skip installing Trio's custom handler, but this "
-        "means MultiErrors will not show full tracebacks.",
-        category=RuntimeWarning,
-    )